Fix: Subscription overwrite (#45)
* fix: overwriting subscriptions

* chore: minor refactor

* chore: replaced deviceId with subscriptionId

* chore: minor naming change

* chore: pr fixes

* chore: refactor and tests

* chore: minor refactor

+++ server/lib/routes/token.js
@@ -1,15 +1,25 @@
 'use strict'
 
+const S = require('fluent-json-schema')
 const uniqid = require('uniqid')
+
+const bodySchema = S.object()
+  .prop('subscriptionId', S.string().required())
+  .prop('secretId', S.string().required())
+
+const schema = {
+  body: bodySchema
+}
 
 async function tokenRoutes(server) {
   server.route({
     method: 'PUT',
-    url: '/api/token/:secretId',
+    url: '/api/token',
+    schema,
     preHandler: server.auth([server.authenticate]),
     handler: async (request, reply) => {
       const { firebaseAdmin } = server
-      const { secretId } = request.params
+      const { subscriptionId, secretId } = request.body
 
       const db = firebaseAdmin.firestore()
 
@@ -20,7 +30,7 @@ async function tokenRoutes(server) {
         .doc(secretId)
         .set({
           token,
-          userId: request.user
+          subscriptionId
         })
       reply.send({ token })
     }
@@ -36,12 +46,37 @@ async function tokenRoutes(server) {
 
       const db = firebaseAdmin.firestore()
 
+      const secretRef = await db
+        .collection('tokens')
+        .doc(secretId)
+        .get()
+
+      if (!secretRef.exists) {
+        return reply.code(404).send('Secret not found')
+      }
+
+      const subscriptionId = secretRef.get('subscriptionId')
+
+      const subscriptionRef = await db
+        .collection('subscriptions')
+        .where(
+          firebaseAdmin.firestore.FieldPath.documentId(),
+          '==',
+          subscriptionId
+        )
+        .where('userId', '==', request.user)
+        .get()
+
+      if (subscriptionRef.empty) {
+        return reply.code(403).send('Not authorized')
+      }
+
       await db
         .collection('tokens')
         .doc(secretId)
         .delete()
 
-      return reply.code(204).send()
+      reply.code(204).send()
     }
   })
 }
